+++ src/lib/index.ts
@@ -18,6 +18,7 @@ export * from "./password";
 export * from "./range";
 export * from "./select";
 export * from "./space";
+export * from "./tabs";
 export * from "./textarea";
 export * from "./theme-toggle";
 export * from "./toast";